fastapi-ithkuil-lodesman/requirements.txt
fastapi
pydantic
uvicorn[standard]
firebase_admin
pyjwt
passlib[bcrypt]